+++ common/hedera-common-nativesupport/src/testFixtures/java/com/hedera/common/testfixtures/rng/RngExtension.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hedera.common.testfixtures;
+package com.hedera.common.testfixtures.rng;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
