+++ cryptography/hedera-cryptography-bls/src/test/java/com/hedera/cryptography/bls/SignaturesLibraryTest.java
@@ -18,7 +18,7 @@ package com.hedera.cryptography.bls;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.hedera.common.testfixtures.WithRng;
+import com.hedera.common.testfixtures.rng.WithRng;
 import com.hedera.cryptography.pairings.api.Curve;
 import com.hedera.cryptography.pairings.api.PairingFriendlyCurves;
 import com.hedera.cryptography.pairings.api.curves.KnownCurves;
